// src/types/muestra.ts
export interface Muestra {
    ID_MUESTRA: number;
    TIPO_MUESTRA_ID: number;
    ID_GENERAL: number;
    ID_PARTICULAR: number;
    ANNO: number;
    TIPO_ANALISIS_ID: number;
    ANALISIS_MODIFICADO?: number;
    FECHA_MUESTREO: Date;
    ENTIDAD_MUESTREO_ID: number;
    DETALLE_MUESTREO?: string;
    OBSERVACIONES?: string;
    FECHA_RECEPCION: Date;
    HORA_RECEPCION?: string;
    EMPLEADO_ID: number;
    FORMATO_ID: number;
    ENTIDAD_ENTREGA_ID: number;
    DETALLE_ENTREGA?: string;
    OBSERVACIONES_ENTREGA?: string;
    CLIENTE_ID: number;
    CENTRO_ID: number;
    REPLACEMENT_ID: number;
    REFERENCIA_CLIENTE: string;
    PRECIO?: number;
    PRECIO_FIJADO?: number;
    FECHA_PREV_FIN: Date;
    OBSERVACIONES_MUESTREO?: string;
    ANULADA: number;
    PRECINTO?: string;
    BANO_ID: number;
    TIPO_FRECUENCIA_ID: number;
    IPA: number;
    INFORME_MANUAL: number;
    FECHA_RECOGIDA?: Date;
    FECHA_COMIENZO?: Date;
    FECHA_FINALIZACION?: Date;
    FECHA_CIERRE?: Date;
    HORA_CIERRE?: string;
    FECHA_ENVIO?: Date;
    MOTIVO_RETRASO_ID?: number;
    RETRASO_OBSERVACION?: string;
    CERRADA: number;
    REVISION_USUARIO?: number;
    CERRADA_USUARIO?: number;
    DOCUMENTO_PAGO?: number;
    IMPRESA?: number;
    ENVIADO_CORREO?: number;
    ULT_EDICION_IMP?: number;
    OP_VUELO?: number;
    OP_INSITU?: number;
    OP_LABMOVIL?: number;
    OP_NORUTINARIA?: number;
    OP_REPETICION?: number;
    SITUACION?: number;
    FIRMA?: number;
    PRODUCTO?: string;
    RESPONSABLE_ID?: number;
    PEDIDO_ID?: number;
    OFERTA_ID?: number;
    PAQUETE_ID?: number;
    ENAC: number;
    NADCAP: number;
    URGENTE: number;
    AJUSTE?: number;
    CONSULTA?: number;
    CONSULTA_OBSERVACIONES?: string;
    AGRUPADA?: number;
    AGRUPADA_MUESTRA_ID?: number;
} 